% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_skip_bioc_nstf.R
\name{check_covr_skip_Bioc_nstf}
\alias{check_covr_skip_Bioc_nstf}
\title{Check for skipped or errored test files (BiocGenerics)}
\usage{
check_covr_skip_Bioc_nstf(pkg_source_path, mapping, test_path)
}
\arguments{
\item{pkg_source_path}{Path to the root of the package source.}

\item{mapping}{A data frame with a column `test_file` listing test file names.}

\item{test_path}{directory}
}
\value{
A data frame listing test files and any issues encountered.
}
\description{
This internal function attempts to run test files mapped to source files and
identifies any that produce warnings or errors. It does not handle `testit`.
}
\keyword{internal}
