% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_report.R
\name{convert_number_to_percent}
\alias{convert_number_to_percent}
\title{Convert number to percent}
\usage{
convert_number_to_percent(value)
}
\arguments{
\item{value}{- number to be converted}
}
\description{
Convert number to percent
}
\keyword{internal}
