% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_skip_stf.R
\name{create_coverage_skip_stf}
\alias{create_coverage_skip_stf}
\title{Run covr with failing tests}
\usage{
create_coverage_skip_stf(
  pkg_source_path,
  pkg,
  test_path = "tests/testthat",
  test_skip,
  test_map,
  preserve_vars = "old"
)
}
\arguments{
\item{pkg_source_path}{Character. path to the package}

\item{pkg}{Character. Name of the package being tested.}

\item{test_path}{character, path to test directory, Default: 'tests/testthat'}

\item{test_skip}{data.frame, mapping of failing tests}

\item{test_map}{data.frame, mapping of all the tests in the package}

\item{preserve_vars}{Character. preserve `old` for CRAN examples}
}
\description{
Run \code{\link[covr]{package_coverage}} ignoring failing tests from \code{\link[testthat]{test_dir}}.
}
\details{
If test_skip is NULL then a new test_dir will be run internally checking
for failing tests.
}
\seealso{
\code{\link[testthat]{test_dir}}
 \code{\link[covr]{package_coverage}}

Other utility: 
\code{\link{map_test}()},
\code{\link{map_tests_stf}()}
}
\concept{utility}
