% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_report.R
\name{create_long_summary_details}
\alias{create_long_summary_details}
\title{Generate Long Summary Details}
\usage{
create_long_summary_details(df)
}
\arguments{
\item{df}{- input data}
}
\description{
Generates the  Long Summary table for the HTML report.
}
\keyword{internal}
