% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_short_path}
\alias{extract_short_path}
\title{Helper to extract "R/<file>" from any path by taking the last two components}
\usage{
extract_short_path(long_file_name)
}
\arguments{
\item{long_file_name}{A string containing the full file path (supports '/' or '\')}
}
\value{
A character string composed of the last two path components, e.g., "R/add.R"
}
\description{
Helper to extract "R/<file>" from any path by taking the last two components
}
