% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_report.R
\name{generate_test_report}
\alias{generate_test_report}
\title{Generate HTML Report for Package Test Assessment}
\usage{
generate_test_report(test_results, output_dir = NULL)
}
\arguments{
\item{test_results}{List containing the results from get_package_coverage function.}

\item{output_dir}{Character string indicating the directory where the report will be saved.}
}
\value{
Path to the generated HTML report.
}
\description{
Generates an HTML report for the package test assessment results using rmarkdown.
}
\examples{
\dontrun{
test_results <- get_package_coverage()
generate_test_report(test_results, output_dir = "path/to/save/report")
}
}
