% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_skip_stf.R
\name{get_stf_test_path}
\alias{get_stf_test_path}
\title{Determine the appropriate standard test path for a package}
\usage{
get_stf_test_path(test_pkg_data, testdir)
}
\arguments{
\item{test_pkg_data}{A named list containing logical flags such as `has_testthat`, `has_testit`, `has_tests_base`, etc.}

\item{testdir}{A character string specifying the root directory where test folders are located.}
}
\value{
A character string representing the path to the test directory, or `NULL` if no known test framework is found.
}
\description{
This internal utility function inspects the `test_pkg_data` structure to identify
which testing framework is used in the package and returns the corresponding test path.
}
\keyword{internal}
