% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_modes.R
\name{get_test_framework_type}
\alias{get_test_framework_type}
\title{Get human-readable test framework type label}
\usage{
get_test_framework_type(test_pkg_data, framework_id)
}
\arguments{
\item{test_pkg_data}{List from \code{check_pkg_tests_and_snaps}.}

\item{framework_id}{Character. Framework ID from \code{detect_frameworks}.}
}
\value{
Character. Label for test_framework_type.
}
\description{
Get human-readable test framework type label
}
\keyword{internal}
