% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_test_metadata}
\alias{get_test_metadata}
\title{Get test metadata}
\usage{
get_test_metadata(executor = NULL)
}
\arguments{
\item{executor}{- user who executes the riskmetrics process 

adapted from mrgvalprep::get_sys_info() and mpn.scorecard}
}
\value{
list with metadata
}
\description{
Get test metadata
}
\keyword{internal}
