% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests_long_sum_stf.R
\name{get_tests_long_sum_stf}
\alias{get_tests_long_sum_stf}
\title{Summarize detailed testthat results}
\usage{
get_tests_long_sum_stf(x)
}
\arguments{
\item{x}{list of test result objects}
}
\value{
data.frame with detailed test results including status and source line links
}
\description{
Internal helper function to generate a long-format summary of testthat results, including status and line references.
}
\keyword{internal}
