% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_report.R
\name{handle_null}
\alias{handle_null}
\title{Helper function to replace NULL with "N/A"}
\usage{
handle_null(x)
}
\arguments{
\item{x}{- input value}
}
\description{
Helper function to replace NULL with "N/A"
}
\keyword{internal}
