% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_modes.R
\name{load_dependencies_into_env}
\alias{load_dependencies_into_env}
\title{Load Imported and Suggested Packages into Environment}
\usage{
load_dependencies_into_env(pkg_source_path, env)
}
\arguments{
\item{pkg_source_path}{Path to the source directory of the package.}

\item{env}{Environment into which the packages should be loaded.}
}
\value{
Invisibly returns TRUE if all packages are loaded successfully.
}
\description{
This function reads the DESCRIPTION file of a package and loads all packages
listed under `Imports` and `Suggests` into the specified environment.
This ensures that functions from those packages are available when running
coverage with `covr::environment_coverage()`.
}
\keyword{internal}
