% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_modes.R
\name{load_package_datasets}
\alias{load_package_datasets}
\title{Load Datasets from a Package}
\usage{
load_package_datasets(pkg_name, env = cov_env)
}
\arguments{
\item{pkg_name}{A character string specifying the name of the package.}

\item{env}{A character string specifying the covr env}
}
\value{
A character vector of dataset names that were attempted to be loaded.
Returns `NULL` if the package is not installed.
}
\description{
Loads all datasets from a specified installed package into the global environment.
It first attempts to load `.rda` files from the package's `data/` directory.
If none are found, it falls back to using `data()` to load datasets listed in the package metadata.
}
\details{
This function is intended for internal use. It silently attempts to load each dataset
and suppresses errors if loading fails. Datasets are loaded into the global environment.
}
\keyword{internal}
