% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tests_stf.R
\name{map_tests_stf}
\alias{map_tests_stf}
\title{Map the hierarchy structure of tests directory in standard testing framework}
\usage{
map_tests_stf(path = "tests/testthat")
}
\arguments{
\item{path}{character, path to tests, Default: 'tests/testthat'}
}
\value{
data.frame
}
\description{
Query a test directory for the unit test structure.
}
\details{
Return data.frame containing which expectations are in the testthat directory
by file/context/description/test/expectation/linerange
}
\seealso{
\code{\link[base]{names}}, \code{\link[base]{lapply}}, \code{\link[base]{rbind}}

Other utility: 
\code{\link{create_coverage_skip_stf}()},
\code{\link{map_test}()}
}
\concept{utility}
