% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_skip_bioc_nstf.R
\name{prepare_for_bioc_runit}
\alias{prepare_for_bioc_runit}
\title{Prepare a BioC/RUnit package for testing}
\usage{
prepare_for_bioc_runit(pkg_name, pkg_source_path)
}
\description{
Detaches common BioC infrastructure packages (if attached) and load_all() the target
package before RUnit tests. Also clears R_TESTS to avoid interference from the check harness.
}
\keyword{internal}
