% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_skip_dt_nstf.R
\name{run_covr_skip_dt_nstf}
\alias{run_covr_skip_dt_nstf}
\title{Run Coverage Analysis on data.table Package Source and Tests}
\usage{
run_covr_skip_dt_nstf(pkg_source_path, test_pkg_data, cov_env)
}
\arguments{
\item{pkg_source_path}{Character string. Path to the root of the package source directory.}

\item{test_pkg_data}{List. Output from `check_pkg_tests_and_snaps()` indicating test framework presence.}

\item{cov_env}{Environment that holds the temporary coverage workspace and
shared state for the run.}
}
\value{
returns a list with coverage object and source files with no tests.
}
\description{
This function maps source files to test files, checks for skipped/problematic tests,
runs the remaining test files manually, calculates coverage, and optionally generates a report.
}
\keyword{internal}
