% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_modes.R
\name{run_one_framework}
\alias{run_one_framework}
\title{Run coverage for a single test framework}
\usage{
run_one_framework(framework, pkg, test_pkg_data, pkg_source_path, covr_timeout)
}
\arguments{
\item{framework}{Character. Framework ID.}

\item{pkg}{Character. Package name.}

\item{test_pkg_data}{List from \code{check_pkg_tests_and_snaps}.}

\item{pkg_source_path}{Character. Path to package source.}

\item{covr_timeout}{Numeric. Timeout for coverage.}
}
\value{
List with total_cov, res_cov, test_framework_type, framework_id.
}
\description{
Run coverage for a single test framework
}
\keyword{internal}
