% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_report.R
\name{safe_value}
\alias{safe_value}
\title{Helper to conditionally apply handle_null or abbreviation}
\usage{
safe_value(x)
}
\arguments{
\item{x}{- value}
}
\description{
Helper to conditionally apply handle_null or abbreviation
}
\keyword{internal}
