% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tests_stf.R
\name{unrowname}
\alias{unrowname}
\title{Add identifier column and remove row names from a data.frame}
\usage{
unrowname(identifier, ret, label)
}
\arguments{
\item{identifier}{character, the name or label to assign to the new column}

\item{ret}{list of data.frames, typically the result of applying a function over multiple files}

\item{label}{character, name of the new column to store the identifier}
}
\value{
data.frame with identifier column prepended and row names removed
}
\description{
Internal helper function to restructure a data.frame by adding an identifier column and removing row names.
}
\keyword{internal}
