% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_skip_stf.R
\name{create_results_summary}
\alias{create_results_summary}
\title{Create a results summary for easier access to scores}
\usage{
create_results_summary(cvr)
}
\arguments{
\item{cvr}{- covr object with full results of skip coverage}
}
\value{
- dataframe with total scores
}
\description{
- retrieve overall percentage and file coverage percentages
}
\keyword{internal}
