% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_report.R
\name{generate_test_details}
\alias{generate_test_details}
\title{Generate Test Details}
\usage{
generate_test_details(assessment_results)
}
\arguments{
\item{assessment_results}{- input data}
}
\description{
Generates the Test Details table for the HTML report.
}
\keyword{internal}
