% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_coverage.R
\name{get_package_coverage}
\alias{get_package_coverage}
\title{get test package coverage}
\usage{
get_package_coverage(path = NULL)
}
\arguments{
\item{path}{(optional) path of locally stored package source code}
}
\value{
list containing results - list containing package coverage
}
\description{
simplified input to assess package for test coverage
}
\examples{
\dontrun{
package_coverage <- get_package_coverage()

package_coverage <- get_package_coverage(path/to/package.tar.gz)
}
}
