% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_skip_nstf.R
\name{get_source_test_mapping_nstf}
\alias{get_source_test_mapping_nstf}
\title{Map R Source Files to Corresponding Test Files}
\usage{
get_source_test_mapping_nstf(pkg_source_path, test_dir)
}
\arguments{
\item{pkg_source_path}{Character string. Path to the root of the package source directory.}

\item{test_dir}{character string. directory where tests are}
}
\value{
A data frame with two columns:
\describe{
  \item{source_file}{Name of the source file in the `R/` directory.}
  \item{test_file}{Name of the matched test file in the `tests/` directory, or \code{NA} if no match is found.}
}
}
\description{
This internal helper function scans the `R/` and `tests/` directories of a package
and attempts to match each source file with a corresponding test file based on filename similarity.
}
\keyword{internal}
