% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package_local.R
\name{install_package_local}
\alias{install_package_local}
\title{Install package locally}
\usage{
install_package_local(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{- source path for install local}
}
\value{
logical - package_installed
}
\description{
Install package locally
}
\examples{
\dontrun{
results <- install_package_local("pkg_source_path")
print(results)
}
}
