% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tests_stf.R
\name{nest_test}
\alias{nest_test}
\title{Nest parsed testthat code into logical test blocks}
\usage{
nest_test(x, token_text = "^context$")
}
\arguments{
\item{x}{data.frame, parse data from utils::getParseData}

\item{token_text}{character, regex pattern to match test block function names, Default: '^context$'}
}
\value{
named list of data.frames, each representing a nested test block
}
\description{
Internal helper function to group parsed testthat code into blocks based on context, test_that, describe, or it calls.
}
\keyword{internal}
