% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_test_functions.R
\name{run_coverage}
\alias{run_coverage}
\title{Run covr and potentially save results to disk}
\usage{
run_coverage(pkg_source_path, timeout = Inf)
}
\arguments{
\item{pkg_source_path}{package installation directory}

\item{timeout}{Timeout to pass to [callr::r_safe()] when running covr.}
}
\value{
list with total coverage and function coverage
}
\description{
Run covr and potentially save results to disk
}
\examples{
\donttest{
r <- getOption("repos")
# save current repo options
old <- options(repos = r)
r["CRAN"] = "http://cran.us.r-project.org"
options(repos = r)

dp <- system.file("test-data", "test.package.0001_0.1.0.tar.gz",
                  package = "test.assessr")

install_list <- set_up_pkg(dp)

package_installed <- install_list$package_installed
pkg_source_path <- install_list$pkg_source_path

package_installed <- install_package_local(pkg_source_path)
package_installed <- TRUE

if (package_installed == TRUE) {
   test_package <- run_coverage(pkg_source_path)
} else {
  message("Package installation failed.")
}
# restore user's repo options
options(old)
}

}
